Move tag pose scripts out of AprilTags example
